Very close to getting signing in and comments working Would just need to work out how to do the scoreboard and finish off comments.
Then finally use bootstrap to fix up the Forms and Scoreboard and interacting with game end stage to grab the score.
May want to put it either with User schema or in its own, the latter would be repetitive but may be required if passport refuses to work with it added.

Also need to quickly write up design challenges and validation (or the lack there of).